requirement.txt
pandas
numpy
seaborn
Flask
python-docx
openpyxl
transformers
torch
tensorflow
-e .